feat(NewsletterSchema): move shared email components to share folder
